Feature/profiles

+++ src/main/java/br/com/vinibelo/githubprofilemanagermutant/repositories/RoleRepository.java
@@ -4,10 +4,8 @@ import br.com.vinibelo.githubprofilemanagermutant.entities.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigInteger;
-
 @Repository
-public interface RoleRepository extends JpaRepository<Role, BigInteger> {
+public interface RoleRepository extends JpaRepository<Role, Long> {
     boolean existsRolesByName(String name);
 
     Role findByName(String name);
